+++ src/IndexFileParser/AsciiIndexFileParser.h
@@ -26,16 +26,16 @@ using std::stringstream;
 class AsciiIndexFileParser: public IndexFileParser {
 private:
   string buffer;
-  size_t buffer_size = 0;
-  size_t buffer_index = 0;
+  u64 buffer_size = 0;
+  u64 buffer_index = 0;
   bool new_read = true;
 
 public:
   AsciiIndexFileParser(
     shared_ptr<ThrowingIfstream> in_stream_,
-    size_t max_indexes_,
-    size_t read_padding_,
-    size_t buffer_size = sixteen_kB
+    u64 max_indexes_,
+    u64 read_padding_,
+    u64 buffer_size = sixteen_kB
   );
   auto generate_batch(
     shared_ptr<IndexesBatch> indexes_batch_,
