+++ src/IndexFileParser/IndexesBeforeNewfileBatchProducer.h
@@ -24,12 +24,12 @@ class IndexesBeforeNewfileBatchProducer:
   friend ContinuousIndexFileParser;
 
 public:
-  explicit IndexesBeforeNewfileBatchProducer(size_t max_batches);
+  explicit IndexesBeforeNewfileBatchProducer(u64 max_batches);
 
 private:
   auto get_default_value() -> shared_ptr<IndexesBeforeNewfileBatch> override;
   auto start_new_batch() -> void;
-  auto add(size_t element) -> void;
+  auto add(u64 element) -> void;
 };
 
 }  // namespace sbwt_search
